Move SetupView to standalone file
